Making more components from too-large View files
